Fix tests
- Check in use test only for darwin
- Terminate test uses for path for darwin (and name for linux)
- Test.app should run on OS X 10.9 (for travis)
