+++ keybase/context_test.go
@@ -104,20 +104,3 @@ func (c testConfigPausedPrompt) keybasePath() string {
 func (c testConfigPausedPrompt) updaterOptions() updater.UpdateOptions {
 	return updater.UpdateOptions{}
 }
-
-func TestContextCheckInUse(t *testing.T) {
-	// In use, force
-	ctx := newContext(&testConfigPausedPrompt{inUse: true, force: true}, log)
-	err := ctx.BeforeApply(updater.Update{})
-	require.NoError(t, err)
-
-	// Not in use
-	ctx = newContext(&testConfigPausedPrompt{inUse: false}, log)
-	err = ctx.BeforeApply(updater.Update{})
-	require.NoError(t, err)
-
-	// In use, user cancels
-	ctx = newContext(&testConfigPausedPrompt{inUse: true, force: false}, log)
-	err = ctx.BeforeApply(updater.Update{})
-	require.EqualError(t, err, "Canceled by user from paused prompt")
-}
