+++ test/Test.app/Contents/Info.plist
@@ -29,7 +29,7 @@
 	<key>DTCompiler</key>
 	<string>com.apple.compilers.llvm.clang.1_0</string>
 	<key>DTPlatformBuild</key>
-	<string>7D175</string>
+	<string>7D1014</string>
 	<key>DTPlatformVersion</key>
 	<string>GM</string>
 	<key>DTSDKBuild</key>
@@ -37,11 +37,11 @@
 	<key>DTSDKName</key>
 	<string>macosx10.11</string>
 	<key>DTXcode</key>
-	<string>0730</string>
+	<string>0731</string>
 	<key>DTXcodeBuild</key>
-	<string>7D175</string>
+	<string>7D1014</string>
 	<key>LSMinimumSystemVersion</key>
-	<string>10.11</string>
+	<string>10.9</string>
 	<key>NSHumanReadableCopyright</key>
 	<string>Copyright © 2016 Keybase. All rights reserved.</string>
 	<key>NSMainNibFile</key>
