+++ test/Test.app/Contents/_CodeSignature/CodeResources
@@ -8,7 +8,7 @@
 		<dict>
 			<key>hash</key>
 			<data>
-			AVhybhV2K3bmuH21nK8afs89io8=
+			d+QK+VLOmER7t5Kt/VCfB5nDBUs=
 			</data>
 			<key>optional</key>
 			<true/>
@@ -20,11 +20,11 @@
 		<dict>
 			<key>hash</key>
 			<data>
-			AVhybhV2K3bmuH21nK8afs89io8=
+			d+QK+VLOmER7t5Kt/VCfB5nDBUs=
 			</data>
 			<key>hash2</key>
 			<data>
-			K/944UN2zuxiLl5RSCyMi3DLph3pz3oNzhETiCiylj0=
+			w4GRCip+qDaMq5FQ7fS9Lls+KdR5g4AbjXGGKuivHXM=
 			</data>
 			<key>optional</key>
 			<true/>
